requirements.txt
Django
django-allauth
django-crispy-forms
whitenoise
django-environ
#psycopg2-binary==2.8.6
######################
#gunicorn


